requirements.txt
fastapi==0.116.1
uvicorn==0.34.2
openai==0.28
openpyxl==3.1.2
pinecone
pydantic==2.10.6
pandas==2.2.2
pytest==8.2.2
python-dotenv==1.0.0